+++ YololEmulator.Tests/Expressions/Num/Modulo.cs
@@ -12,7 +12,7 @@ namespace YololEmulator.Tests.Expressions.Num
 
             var a = result.GetVariable("a");
 
-            Assert.AreEqual(1, a.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
         }
 
         [TestMethod]
@@ -23,8 +23,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(5, a.Value.Number);
-            Assert.AreEqual(2, b.Value.Number);
+            Assert.AreEqual(5, (int)a.Value.Number);
+            Assert.AreEqual(2, (int)b.Value.Number);
         }
 
         [TestMethod]
@@ -35,8 +35,8 @@ namespace YololEmulator.Tests.Expressions.Num
             var a = result.GetVariable("a");
             var b = result.GetVariable("b");
 
-            Assert.AreEqual(3, a.Value.Number);
-            Assert.AreEqual(2, b.Value.Number);
+            Assert.AreEqual(3, (int)a.Value.Number);
+            Assert.AreEqual(2, (int)b.Value.Number);
         }
     }
 }
