+++ YololEmulator.Tests/Expressions/Num/Tangent.cs
@@ -11,7 +11,7 @@ namespace YololEmulator.Tests.Expressions.Num
         {
             var result = TestExecutor.Execute("a = tan(0)");
             var a = result.GetVariable("a");
-            Assert.AreEqual(0, a.Value.Number);
+            Assert.AreEqual(0, (int)a.Value.Number);
         }
 
         [TestMethod]
@@ -19,7 +19,7 @@ namespace YololEmulator.Tests.Expressions.Num
         {
             var result = TestExecutor.Execute("a = tan(45)");
             var a = result.GetVariable("a");
-            Assert.AreEqual(1, a.Value.Number);
+            Assert.AreEqual(1, (int)a.Value.Number);
         }
 
         [TestMethod]
